GH-35069: [Archery][Release] Remove retrieving ARROW issue from migration comment on Archery release (#35070)
### Rationale for this change
This is not required anymore because even when the issue was originally created on JIRA the PR has been merged referring to GH.

### What changes are included in this PR?

Remove retrieving for release curate and release notes the original `ARROW-` prefix issue id for JIRA and use the new `GH-` issue id.

### Are these changes tested?

I have tested it locally.

### Are there any user-facing changes?
No
* Closes: #35069

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>