GH-35062: [Go][CI] Fix verification failures (#35077)
### Rationale for this change
In the conda release verification we saw an error failure due to "build constraints exclude all files" in the `memory/mallocator` directory. This is easily fixed by adding a `doc.go` file with a package documentation and no build constraint on it to the directory. While here, to ensure everything is handled correctly, the `cgo` constraint was added to the default `mallocator` tagged files (and `!cgo` to the defaults) so that it's explicit rather than implicit (by virtue of importing mallocator which requires cgo).

This should, hopefully, also fix the other macos verification failure by ensuring the Time64 nanosecond parsing test doesn't just drop the extraneous 0.

### What changes are included in this PR?

### Are there any user-facing changes?
There shouldn't be.

* Closes: #35062

Authored-by: Matt Topol <zotthewizard@gmail.com>
Signed-off-by: Jacob Wujciak-Jens <jacob@wujciak.de>