GH-35118: [Format][FlightSQL] More use `int32` to refer to 32-bit integers rather than `int` (#35213)
### Rationale for this change

There are more inconsistency of the spec format found  -- see details on the original issue #35118. https://github.com/apache/arrow/pull/35120 is the first PR with the same fix.

### What changes are included in this PR?

Use `int32` to refer to 32-bit integers rather than `int`

### Are these changes tested?

No, only comments are changed

### Are there any user-facing changes?

This clarifies a small corner case in the document

* Closes: #35118

Authored-by: Chunchun <14298407+appletreeisyellow@users.noreply.github.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>