GH-34907: [Docs][R] Version selector reports that release version is dev (#35103)
This PR moves the functionality to select a different version of the docs from a dropdown menu (which replaces the version number span on the page) to a link in the footer, which links to a HTML page with clickable links to the different docs versions.

This resolves the problem in #34907 and also fixes #14969 in which the JavaScript used in the dropdown interferes with the JavaScript code used for the search box (present since the upgrade to BS5).

This change also simplifies the maintenance of the version selection functionality, though does remove the ability to navigate directly to the same reference page in a different version - overall I think this minor loss of functionality is a worthwhile trade for simpler maintenance.

Footer before (note that the author information is already in the sidebar, so this is duplication):
![image](https://user-images.githubusercontent.com/13715823/231714608-73ec3e0b-cc28-4f41-9b22-a87033ab3977.png)

Footer after:
![image](https://user-images.githubusercontent.com/13715823/231714402-e04b1c43-495a-477b-ae1e-4035b35b088b.png)

* Closes: #34907

Lead-authored-by: Nic Crane <thisisnic@gmail.com>
Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>