GH-35410: [CI] Pin urllib3 for gcs testbench (#35412)
### Rationale for this change

gcs-toolbench's latest release currently pins requests-toolbelt to 0.10.0 which is not compatible with urllib >= 2.0.  Until there is a new release of gcs-toolbench I think we need to pin urllib to be < 2.0.

### What changes are included in this PR?

A change to the install_gcs_toolbench.sh script that pins urllib to < 2.

### Are these changes tested?

Yes, CI uses this script today.

### Are there any user-facing changes?

No
* Closes: #35410

Authored-by: Weston Pace <weston.pace@gmail.com>
Signed-off-by: Weston Pace <weston.pace@gmail.com>